bunch of ui + refactor
