+++ src/components/DappBrowser/search-input/TabButton.tsx
@@ -8,16 +8,10 @@ import { BlurView } from '@react-native-community/blur';
 import React, { useCallback, RefObject } from 'react';
 import { TextInput } from 'react-native';
 import { DappBrowserShadows } from '../DappBrowserShadows';
+import { useBrowserContext } from '../BrowserContext';
 
-export const TabButton = ({
-  inputRef,
-  isFocused,
-  toggleTabView,
-}: {
-  inputRef: RefObject<TextInput>;
-  isFocused: boolean;
-  toggleTabView: () => void;
-}) => {
+export const TabButton = ({ toggleTabView }: { toggleTabView: () => void }) => {
+  const { isSearchInputFocused, searchInputRef } = useBrowserContext();
   const { isDarkMode } = useColorMode();
   const fillSecondary = useForegroundColor('fillSecondary');
   const separatorSecondary = useForegroundColor('separatorSecondary');
@@ -27,14 +21,14 @@ export const TabButton = ({
   const buttonColor = IS_IOS ? buttonColorIOS : buttonColorAndroid;
 
   const onPress = useCallback(() => {
-    if (!isFocused) {
+    if (!isSearchInputFocused) {
       // open tabs
       toggleTabView();
     } else {
       // close keyboard
-      inputRef?.current?.blur();
+      searchInputRef?.current?.blur();
     }
-  }, [isFocused, inputRef, toggleTabView]);
+  }, [isSearchInputFocused, searchInputRef, toggleTabView]);
 
   return (
     <DappBrowserShadows>
@@ -42,12 +36,12 @@ export const TabButton = ({
         as={ButtonPressAnimation}
         borderRadius={22}
         onPress={onPress}
-        style={{ height: 44, paddingTop: isFocused ? 1 : undefined, width: 44 }}
+        style={{ height: 44, paddingTop: isSearchInputFocused ? 1 : undefined, width: 44 }}
         alignItems="center"
         justifyContent="center"
       >
-        <Text align="center" color="labelSecondary" size="icon 17px" weight={isFocused ? 'heavy' : 'bold'}>
-          {isFocused ? '􀆈' : '􀐅'}
+        <Text align="center" color="labelSecondary" size="icon 17px" weight={isSearchInputFocused ? 'heavy' : 'bold'}>
+          {isSearchInputFocused ? '􀆈' : '􀐅'}
         </Text>
         {IS_IOS && (
           <Box
