+++ src/components/DappBrowser/search/Search.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { SearchBar } from './SearchBar';
+import { SearchResults } from './SearchResults';
+
+export const Search = () => {
+  return (
+    <>
+      <SearchResults />
+      <SearchBar />
+    </>
+  );
+};
